+++ source/src/MapConfig.cs
@@ -5,7 +5,7 @@ namespace EnhancedBattleTest
     public class MapConfig
     {
         public string MapNameSearchText = String.Empty;
-        public int WallHitPoint = 0;
+        public int BreachedWallCount = 0;
         public int SceneLevel = 1;
         public string Season = String.Empty;
     }